+++ src/components/NavItems/Assistant/AssistantScrapeResults/assistantUtils.jsx
@@ -1,7 +1,5 @@
 import React from "react";
 import _ from "lodash";
-import { ColourGradientTooltipContent } from "./AssistantTextClassification";
-import Tooltip from "@mui/material/Tooltip";
 
 /**
  * Interpolate RGB between an arbitrary range
@@ -85,35 +83,17 @@ function treeMapToElementsRecursive(
   if ("span" in treeElem) {
     const span = treeElem.span;
     if (spanHighlightIndices === null) {
-      console.log("No span highlight: ", text.substring(span.start, span.end));
       childElems.push(text.substring(span.start, span.end));
     } else {
-      console.log("Span highlight: ", text.substring(span.start, span.end));
       let currentIndex = span.start;
       for (let i = 0; i < spanHighlightIndices.length; i++) {
         const hSpan = spanHighlightIndices[i];
-        console.log(
-          "Matching span",
-          span.start,
-          span.end,
-          hSpan.indices[0],
-          hSpan.indices[1],
-        );
         const hSpanStart = hSpan.indices[0];
         const hSpanEnd = hSpan.indices[1];
         if (
           (span.start <= hSpanStart && hSpanStart <= span.end) ||
           (span.start <= hSpanEnd && hSpanEnd <= span.end)
         ) {
-          //If there's an overlap
-          console.log(
-            "Found lapping span ",
-            span.start,
-            span.end,
-            hSpanStart,
-            hSpanEnd,
-          );
-
           // If span doesn't start before the current index
           if (hSpanStart > currentIndex) {
             childElems.push(text.substring(currentIndex, hSpanStart));
@@ -123,7 +103,6 @@ function treeMapToElementsRecursive(
             hSpanStart < span.start ? span.start : hSpanStart;
           const boundedEnd = hSpanEnd > span.end ? span.end : hSpanEnd;
           if (wrapFunc) {
-            console.log("Wrapping: ", text.substring(boundedStart, boundedEnd));
             childElems.push(
               wrapFunc(
                 text.substring(boundedStart, boundedEnd),
@@ -133,10 +112,6 @@ function treeMapToElementsRecursive(
               ),
             );
           } else {
-            // console.log(
-            //   "Not wrapping: ",
-            //   text.substring(boundedStart, boundedEnd),
-            // );
             childElems.push(text.substring(boundedStart, boundedEnd));
           }
